i'll continue to debug after my flutter upgrade.

+++ flutter_app/lib/model.product.dart
@@ -15,7 +15,7 @@ class Product {
     return Product(
       id: json['_id'],
       name: json['name'],
-      price: json['price'].toDouble(),
+      price: json['price'],
       description: json['description'],
     );
   }
